+++ src/Actions/Campaigns/PrepareSubjectAction.php
@@ -9,8 +9,6 @@ class PrepareSubjectAction
 {
     public function execute(Campaign $campaign)
     {
-        $campaign->pullSubjectFromMailable();
-
         $this->replacePlaceholdersInSubject($campaign);
 
         $campaign->save();