+++ src/Components/DateTimeFieldComponent.php
@@ -2,17 +2,17 @@
 
 namespace Spatie\Mailcoach\Components;
 
-use Illuminate\Support\Carbon;
+use Carbon\CarbonInterface;
 use Illuminate\Support\Collection;
 use Illuminate\View\Component;
 
 class DateTimeFieldComponent extends Component
 {
-    public Carbon $value;
+    public CarbonInterface $value;
 
     public string $name;
 
-    public function __construct(string $name, ?Carbon $value = null)
+    public function __construct(string $name, ?CarbonInterface $value = null)
     {
         $this->value = $value ?? now();
         $this->name = $name;
@@ -27,8 +27,8 @@ class DateTimeFieldComponent extends Component
 
     public function minuteOptions(): Collection
     {
-        return collect(range(0, 60, 15))->mapWithKeys(function (int $minutes) {
-            return [$minutes => str_pad((string)$minutes, 2, '0', STR_PAD_LEFT)];
+        return collect(range(0, 59, 15))->mapWithKeys(function (int $minutes) {
+            return [$minutes => str_pad($minutes, 2, '0', STR_PAD_LEFT)];
         });
     }