+++ src/Rules/DateTimeFieldRule.php
@@ -2,8 +2,9 @@
 
 namespace Spatie\Mailcoach\Rules;
 
-use Carbon\Carbon;
+use Carbon\CarbonInterface;
 use Illuminate\Contracts\Validation\Rule;
+use Illuminate\Support\Facades\Date;
 use InvalidArgumentException;
 
 class DateTimeFieldRule implements Rule
@@ -50,13 +51,13 @@ class DateTimeFieldRule implements Rule
         return $this->message;
     }
 
-    public function parseDateTime(array $value): ?Carbon
+    public function parseDateTime(array $value): ?CarbonInterface
     {
         try {
             $hours = str_pad($value['hours'], 2, '0', STR_PAD_LEFT);
             $minutes = str_pad($value['minutes'], 2, '0', STR_PAD_LEFT);
 
-            return Carbon::createFromFormat(
+            return Date::createFromFormat(
                 'Y-m-d H:i',
                 "{$value['date']} {$hours}:{$minutes}"
             );