+++ tests/Jobs/SendCampaignJobTest.php
@@ -230,7 +230,7 @@ class SendCampaignJobTest extends TestCase
 
         $campaign->emailList->update(['campaign_mailer' => 'array']);
 
-        dispatch(new SendCampaignJob($campaign));
+        $campaign->send();
 
         $messages = app(MailManager::class)->mailer('array')->getSwiftMailer()->getTransport()->messages();
 
@@ -251,7 +251,7 @@ class SendCampaignJobTest extends TestCase
 
         $campaign->emailList->update(['campaign_mailer' => 'array']);
 
-        dispatch(new SendCampaignJob($campaign));
+        $campaign->send();
 
         $messages = app(MailManager::class)->mailer('array')->getSwiftMailer()->getTransport()->messages();
 
@@ -278,7 +278,7 @@ class SendCampaignJobTest extends TestCase
 
         $campaign->emailList->update(['campaign_mailer' => 'array']);
 
-        dispatch(new SendCampaignJob($campaign));
+        $campaign->send();
 
         $messages = app(MailManager::class)->mailer('array')->getSwiftMailer()->getTransport()->messages();
 
@@ -302,7 +302,7 @@ class SendCampaignJobTest extends TestCase
 
         $campaign->emailList->update(['campaign_mailer' => 'array']);
 
-        dispatch(new SendCampaignJob($campaign));
+        $campaign->send();
 
         $messages = app(MailManager::class)->mailer('array')->getSwiftMailer()->getTransport()->messages();