+++ tests/Support/ReplacesModelAttributesTest.php
@@ -4,6 +4,7 @@ namespace Spatie\Mailcoach\Tests\Support;
 
 use Spatie\Mailcoach\Models\Subscriber;
 use Spatie\Mailcoach\Support\Replacers\Concerns\ReplacesModelAttributes;
+use Spatie\Mailcoach\Tests\Factories\SubscriberFactory;
 use Spatie\Mailcoach\Tests\TestCase;
 
 class ReplacesModelAttributesTest extends TestCase
@@ -56,6 +57,20 @@ class ReplacesModelAttributesTest extends TestCase
 
         $output = $this->classWithTrait->replaceModelAttributes(
             "This is ::subscriber.non_existing_attribute::",
+            'subscriber',
+            $subscriber
+        );
+
+        $this->assertEquals("This is ", $output);
+    }
+
+    /** @test */
+    public function it_will_not_thrown_an_exception_when_trying_to_replace_a_non_existing_schemaless_attribute()
+    {
+        $subscriber = SubscriberFactory::new()->create();
+
+        $output = $this->classWithTrait->replaceModelAttributes(
+            "This is ::subscriber.extra_attributes.non_existing_attribute::",
             'subscriber',
             $subscriber
         );